requirements.txt
# Formatting
autopep8==1.6.0

# Django & Django REST framework
Django==3.2.2

# Django toolkit
django-cors-headers==3.11.0

# Postgres database
psycopg2==2.9.3

# Loadings .env file
python-dotenv==0.15.0

# Ninja
django-ninja==0.21.0
django-ninja-extra==0.18.7
django-ninja-jwt==5.2.5
orjson==3.8.10

# Image
Pillow==9.0.1

# Facebook
python-facebook-api==0.15.0

requests==2.31.0

PyJWT==2.7.0

# markdown
bs4==0.0.1
beautifulsoup4==4.12.2
Markdown==3.4.3

# Firebase
firebase-admin==6.2.0